requirements.txt
psycopg2-binary>=2.9.9
sqlglot>=25.0.0
numpy>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
openpyxl>=3.1.5
langgraph>=0.2.50
langchain>=0.3.7
langchain-core>=0.3.19
langchain-ollama>=0.2.0
langsmith>=0.1.143
